An attempt at implementing block deserializing logic that turns out to be a dead end. This will be heavily refactored as the next work item.

+++ Minecraft.Client/Orientation.cs
@@ -1,6 +1,6 @@
 ﻿namespace Decent.Minecraft.Client
 {
-    public enum Orientation : ushort
+    public enum Orientation : byte
     {
         UpDown = 0x0,
         EastWest = 0x4,
